Корпоративная программа Spring Framework. Домашнее задание 15. Вынес @Schedule из сервисов в конфиг, отключил его в тестах.
