+++ Homework15/homework-15-library/src/main/java/ru/otus/library/Main.java
@@ -3,10 +3,8 @@ package ru.otus.library;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.scheduling.annotation.EnableScheduling;
 
 @EnableCaching
-@EnableScheduling
 @SpringBootApplication
 public class Main {
 	public static void main(String[] args) {
